Laget KopierLenke som eget komponent laget et lagret samtykkeskjema
