+++ frontend/src/style/less/komponenter/kopierLenke/kopierLenke.less
@@ -1,0 +1,13 @@
+// Hele lenkekomponentet
+.lenke-komponent {
+    display: flex;
+    width: 100%;
+}
+
+// Lenketeksten
+.lenke-input {
+    width: 100%;
+}
+.opprettelse-fullført > .hovedomrade {
+    min-height: 36rem;
+}
