+++ frontend/src/style/less/main.less
@@ -18,6 +18,7 @@
 @import './komponenter/lagSamtykkeskjema/ansvarligPersonerInputs.less';
 @import './komponenter/lagSamtykkeskjema/kontaktPersonerInputs.less';
 @import './komponenter/lagSamtykkeskjema/jaNeiValg.less';
+@import './komponenter/kopierLenke/kopierLenke.less';
 
 //CONTAINERE
 @import './containere/brukerKnappToppBar.less';
@@ -34,6 +35,7 @@
 
 @import './visninger/admin/opprettSamtykkeskjema.less';
 @import './visninger/admin/opprettelseFullført.less';
+@import './visninger/admin/mineSamtykkeskjemaer.less';
 
 @import './visninger/landingsside/landingsside.less';
