+++ frontend/src/style/less/visninger/admin/opprettSamtykkeskjema.less
@@ -1,4 +1,4 @@
-//Hele visningen
+// Hele visningen
 .landingsside-opprett {
     height: 100%;
     display: flex;
@@ -6,14 +6,14 @@
     align-items: center;
 }
 
-//Undertittel
+// Undertittel
 .typo-systemtittel.skjema-info {
     margin-top: 4%;
     margin-bottom: 8%;
     white-space: nowrap;
 }
 
-//knapper
+// Knapper
 .knapper {
     height: 11rem;
     display: flex;
@@ -22,6 +22,7 @@
     text-decoration: none;
 }
 
+// Hovedomrade
 .landingsside-opprett > .hovedomrade {
     min-height: 36rem;
 }
