+++ frontend/src/style/less/visninger/admin/opprettelseFullført.less
@@ -6,17 +6,6 @@
     align-items: center;
 }
 
-// Hele lenkekomponentet
 .lenke-komponent {
-    display: flex;
-    width: 100%;
     margin: 1rem 0 3rem;
-}
-
-// Lenketeksten
-.lenke-input {
-    width: 100%;
-}
-.opprettelse-fullført > .hovedomrade {
-    min-height: 36rem;
 }
